requirements.txt
# MkDocs and theme
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocs-redirects>=1.2.0

# Extensions
pymdown-extensions>=10.0

# Aggregation script dependencies
pyyaml>=6.0
aiohttp>=3.9.0
pydantic>=2.0.0

# Schema validation
check-jsonschema>=0.28.0

# Template management (ADR-007)
ruamel.yaml>=0.18.0
jsonschema>=4.20.0

# Development
pre-commit>=3.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
